Refactor test case methods in test_helper.py
Refactored a few lines in the `test_helper.py` file to make the code clearer and less verbose. The change mainly involves formatting the lines so no multi-line statements are used, and adding a new method `set_defaults(call_)` that assign default values for the summary and text attributes of the call.

[release:minor]